+++ src/composables/useDark.ts
@@ -1,0 +1,6 @@
+export const useDarkProps = {
+  dark: {
+    type: Boolean,
+    default: null
+  }
+};