+++ src/packages/components.ts
@@ -1,2 +1,3 @@
 export * from './button';
 export * from './icons';
+export * from './loading';