+++ src/packages/divider/__test__/index.spec.ts
@@ -1,0 +1,11 @@
+import { mount } from '@vue/test-utils';
+import EDivider from '..';
+
+describe('EDivider', () => {
+
+  it('EDivider snapshot', () => {
+    const wrapper = mount(EDivider);
+    expect(wrapper.html()).toMatchSnapshot();
+    wrapper.unmount();
+  });
+});