+++ src/packages/divider/demo/demo0.vue
@@ -1,0 +1,7 @@
+<template>
+  <EDivider />
+</template>
+
+<script  setup>
+import { EDivider } from 'eurus-ui';
+</script>