+++ src/packages/icons/src/index.tsx
@@ -2,7 +2,7 @@ import type { PropType } from 'vue';
 import { renderSlot, h, computed, defineComponent } from 'vue';
 
 import IconList from './iconList';
-import { useSize } from '@/packages/_shared/useSize';
+import { useSize } from '@/composables/useSize';
 
 const IconProps = {
   size: [Number, String],