+++ src/packages/loading/demo/demo0.vue
@@ -1,0 +1,10 @@
+<template>
+  <ELoading v-model:show="show" />
+</template>
+
+<script  setup>
+import { ELoading } from 'eurus-ui';
+
+import { ref } from 'vue';
+const show = ref(false);
+</script>