+++ CHANGELOG.md
@@ -5,3 +5,60 @@ toc: false
 docClass: timeline
 ---
 
+## [0.0.4](https://github.com/jiangxd2016/eurus-ui/compare/1a3959f22417636dde8cf1723440d8d938eb91fd...v0.0.4) `2022-07-07`
+
+
+### Bug Fixes
+
+* **playground:** fix icon ([0efb90b](https://github.com/jiangxd2016/eurus-ui/commit/0efb90b5e349693e8cb983de8f98f7ec8f2df8b7))
+
+
+### Features
+
+* **1:** :sparkles: 2 ([1a3959f](https://github.com/jiangxd2016/eurus-ui/commit/1a3959f22417636dde8cf1723440d8d938eb91fd))
+* docs theme ([fd2df4f](https://github.com/jiangxd2016/eurus-ui/commit/fd2df4ffe16b325d49f13f0fc653e8c74a55816e))
+* **script:** [WIP] gen template ([ea75ff1](https://github.com/jiangxd2016/eurus-ui/commit/ea75ff16e3a0e93303133de4a6bb638b30c2042d))
+
+
+### BREAKING CHANGES
+
+* **1:** 4
+
+## [0.0.5](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.4...v0.0.5) `2022-07-07`
+
+## [0.0.6](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.5...v0.0.6) `2022-07-07`
+
+## [0.0.7](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.6...v0.0.7) `2022-07-07`
+
+## [0.0.8](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.7...v0.0.8) `2022-07-07`
+
+## [0.0.10](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.8...v0.0.10) `2022-07-07`
+
+## [0.0.11](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.10...v0.0.11) `2022-07-07`
+
+## [0.0.12](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.11...v0.0.12) `2022-07-07`
+
+## [0.0.13](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.12...v0.0.13) `2022-07-07`
+
+## [0.0.14](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.13...v0.0.14) `2022-07-07`
+
+## [0.0.15](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.14...v0.0.15) `2022-07-07`
+
+## [0.0.16](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.15...v0.0.16) `2022-07-07`
+
+## [0.0.18](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.16...v0.0.18) `2022-07-07`
+
+## [0.0.19](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.18...v0.0.19) `2022-07-07`
+
+## [0.0.20](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.19...v0.0.20) `2022-07-07`
+
+## [0.0.1](https://github.com/jiangxd2016/eurus-ui/compare/v0.0.20...v0.0.1) `2022-07-08`
+
+
+### Features
+
+* **components:**  update icons ([5a83413](https://github.com/jiangxd2016/eurus-ui/commit/5a834130641c6dfb1a6d080e7985ac306e76e0c5))
+
+
+
+